refactor(blorktools): landing-screen to page
