+++ packages/blorktools/src/asset_debugger/landing-page/dropzone-preview.js
